+++ OneLogin.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -131,8 +131,8 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/govuk-one-login/mobile-id-check-ios-sdk.git",
       "state" : {
-        "revision" : "e7c7a8a1a680dfc9ba7740b229e3b35238dec155",
-        "version" : "8.3.4"
+        "branch" : "fix/dcmaw-13381-readid-landscape",
+        "revision" : "e9ab51fc86b76a94666671b4b5a6bcfc0b1c8505"
       }
     },
     {
