+++ Sources/Qualification/QualifyingCoordinator.swift
@@ -1,6 +1,5 @@
 import Authentication
 import Coordination
-import DocumentCheck
 import GDSCommon
 import Logging
 import Networking
@@ -195,62 +194,6 @@ extension QualifyingCoordinator {
                     self.deeplink = nil
                 }
             }
-        }
-    }
-}
-
-public typealias IDCheckNavigationController = ConfirmNavigationController
-
-class CustomTabBarController: UITabBarController {
-    override open var shouldAutorotate: Bool {
-        get {
-            if let vc = self.selectedViewController?.presentedViewController, vc is IDCheckNavigationController {
-                return vc.shouldAutorotate
-            } else {
-                return super.shouldAutorotate
-            }
-        }
-        set {
-            // Empty implementation
-        }
-    }
-    
-    override open var preferredInterfaceOrientationForPresentation: UIInterfaceOrientation {
-        get {
-            if let vc = self.selectedViewController?.presentedViewController, vc is IDCheckNavigationController {
-                return vc.preferredInterfaceOrientationForPresentation
-            } else {
-                return super.preferredInterfaceOrientationForPresentation
-            }
-        }
-        set {
-            // Empty implementation
-        }
-    }
-    
-    override open var supportedInterfaceOrientations: UIInterfaceOrientationMask {
-        get {
-            if let vc = self.selectedViewController?.presentedViewController, vc is IDCheckNavigationController {
-                return vc.supportedInterfaceOrientations
-            } else {
-                return super.supportedInterfaceOrientations
-            }
-        }
-        set {
-            // Empty implementation
-        }
-    }
-    
-    override open var preferredStatusBarStyle: UIStatusBarStyle {
-        get {
-            if let vc = self.selectedViewController?.presentedViewController, vc is IDCheckNavigationController {
-                return vc.preferredStatusBarStyle
-            } else {
-                return super.preferredStatusBarStyle
-            }
-        }
-        set {
-            // Empty implementation
         }
     }
 }
